requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.2
dearpygui==1.8.0
defusedxml==0.7.1
Django==4.1.5
django-debug-toolbar==4.1.0
django-filter==22.1
djangorestframework==3.14.0
export==0.2.1
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
protobuf==4.24.1
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
